Развитие архитектуры
Базовая ViewModel для карточек, их идентификация и универсальная отрисовка

+++ app/src/main/java/com/example/testapp/domain/EnabledCard.kt
@@ -1,3 +1,3 @@
 package com.example.testapp.domain
 
-data class EnabledCard(val type: Cards, val id: Int)
+data class EnabledCard(val type: Cards, val id: Long)
